Updated download and install section, edit .kitchen.yml section, and added .kitchen.yml example file.
This branch includes changes after testing the pipeline cookbook instructions. I updated the install section to include the prerequisits of installing test-kitchen and berkshelf.  I consolidated the .kitchen.yml edit section (it was two separate sections) to one section, and added centos-6.5 to the platforms (original specifed running centos-6.5 but did not show an example of adding the new platform). I also included an example .kitchen.yml for reference.